comit de permisos

+++ vistas/mostrar_pacientes_vista.php
@@ -14,6 +14,20 @@ $sql2="INSERT  INTO TBL_BITACORA (BIT_CODIGO,USU_CODIGO,OBJ_CODIGO,BIT_ACCION,BI
 VALUES (:id,:usuc,:objeto,:accion,:descr,:fecha)";
 $resultado2=$conexion->prepare($sql2);	
 $resultado2->execute(array(":id"=>NULL,":usuc"=>$_SESSION["id_us"],":objeto"=>11,":accion"=>'CONSULTA',":descr"=>'MUESTRA LA LISTA DE USUARIOS  QUE HAY MANTENIMIENTO',":fecha"=>date("Y-m-d H:m:s")));
+
+$ROL = $_SESSION['ROL'];
+$_SESSION['PANTALLA'] = 17;
+$PANTALLA = $_SESSION['PANTALLA'];
+$sql3 = "select * from tbl_permisos where ROL_CODIGO = :rol and OBJ_CODIGO = :pantalla ";
+$resultado3=$conexion->prepare($sql3);	
+$resultado3->execute(array(":rol"=>$ROL,":pantalla"=>$PANTALLA));
+$DATOS = $resultado3->fetch(PDO::FETCH_ASSOC);
+ $CONSULTAR = $DATOS['PERM_CONSULTAR'];
+ $INSERTAR = $DATOS['PERM_INSERTAR'];
+ $ELIMINAR = $DATOS['PERM_ELIMINAR'];
+ $ACTUALIZAR = $DATOS['PERM_ACTUALIZAR'];
+ $PERM_OBJ = $DATOS['PERM_OBJ'];
+
 ?>
 <!DOCTYPE html>
 <html>
@@ -85,7 +99,6 @@ $resultado2->execute(array(":id"=>NULL,":usuc"=>$_SESSION["id_us"],":objeto"=>11
         </div>
         <div class="pull-left info">
           <p><?php echo $_SESSION["usu"];?></p>
-        
           </div>
       </div>
       <!-- search form -->
@@ -102,12 +115,13 @@ $resultado2->execute(array(":id"=>NULL,":usuc"=>$_SESSION["id_us"],":objeto"=>11
           <span>Usuarios</span>
         </a>
         <!-- subtitulos de Usuario -->
-        <ul class="treeview-menu">
-          <li><a href="../vistas/insertar_mant_vista.php"><i class="fa fa-plus-square"></i>Añadir Usuarios</a></li>
-          <li><a href="../vistas/mostrar_vista.php"><i class="fa fa-list"></i>Lista de Usuarios</a></li>
-         
 
+        <?php if ($CONSULTAR == 1){ ?>
+          <ul class="treeview-menu">
+          <li><a href="../vistas/mostrar_vista.php"><i class="fa fa-list"></i>Lista de Usuarios</a></li>
         </ul>
+       <?php } ?>
+        
       </li>
        <!-- Titulo de Empleados -->
       <li class="treeview">
@@ -117,11 +131,13 @@ $resultado2->execute(array(":id"=>NULL,":usuc"=>$_SESSION["id_us"],":objeto"=>11
 
         </a>
         <!-- subtitulos de Empleados -->
-        <ul class="treeview-menu">
-        <li><a href="../vistas/insertar_empleado_vista.php"><i class="fa fa-plus-square"></i>Añadir Empleado</a></li>
-          <li><a href="../vistas/mostrar_empleados_vista.php"><i class="fa fa-list"></i> Lista de Empleados</a></li>
 
+        <?php if ($CONSULTAR == 1){ ?>
+          <ul class="treeview-menu">
+          <li><a href="../vistas/mostrar_empleados_vista.php"><i class="fa fa-list"></i> Lista de Empleados</a></li>
         </ul>
+        <?php } ?>
+        
       </li>
      
       <!-- Titulo de Pacientes -->
@@ -132,27 +148,42 @@ $resultado2->execute(array(":id"=>NULL,":usuc"=>$_SESSION["id_us"],":objeto"=>11
 
         </a>
         <!-- subtitulos de Pacientes -->
-        <ul class="treeview-menu">
-          <li><a href="../vistas/insertar_pacientes_vista.php"><i class="fa fa-plus-square"></i>Añadir Paciente</a></li>
+        <?php if ($PERM_OBJ == 1){ ?>
+          <ul class="treeview-menu">
           <li><a href="../vistas/mostrar_pacientes_vista.php"><i class="fa fa-list"></i>Lista de Pacientes</a></li>
-          
         </ul>
+       <?php } ?>
+        
       </li>
       <!-- Titulo de Expedientes -->
       <li class="treeview">
         <a href="#">
           <i class="fa fa-folder-open-o"></i>
-          <span>Expedientes</span>
+          <span>Expedientes Nutricionista</span>
 
           </a>
         <!-- subtitulos de Expedientes -->
-        <ul class="treeview-menu">
-        <li><a href="../vistas/insertar_expedienten_vista.php"><i class="fa fa-plus-square"></i>Expediente Nutricional</a></li>
+        <?php if ($PERM_OBJ == 1){ ?>
+          <ul class="treeview-menu">
           <li><a href="../vistas/mostrar_expedienten_vista.php"><i class="fa fa-list"></i>Mostrar Expediente Nutricional</a></li>
-          <li><a href="../vistas/insertar_expediented_vista.php"><i class="fa fa-plus-square"></i>Expediente Médico </a></li>
-          <li><a href="../vistas/mostrar_expediented_vista.php"><i class="fa fa fa-list"></i>Mostrar Expediente Médico </a></li>
-
         </ul>
+         <?php } ?>
+        
+      </li>
+
+      <li class="treeview">
+        <a href="#">
+          <i class="fa fa-folder-open-o"></i>
+          <span>Expedientes Medico </span>
+
+          </a>
+        <!-- subtitulos de Expedientes -->
+        <?php if ($PERM_OBJ == 1){ ?>
+          <ul class="treeview-menu">
+          <li><a href="../vistas/mostrar_expediented_vista.php"><i class="fa fa fa-list"></i>Mostrar Expediente Médico </a></li>
+        </ul>
+        <?php } ?>
+       
       </li>
       <!-- Titulo de Citas -->
       <li class="treeview">
@@ -160,16 +191,20 @@ $resultado2->execute(array(":id"=>NULL,":usuc"=>$_SESSION["id_us"],":objeto"=>11
           <i class="fa fa-calendar"></i>
           <span>Citas</span>
           </a>
-        <ul class="treeview-menu">
-          <li><a href="../vistas/insertar_cita_vista.php"><i class="fa fa-plus-square"></i>Añadir cita</a></li>
-          <li><a href="../vistas/mostrar_citas_vista.php"><i class="fa fa-list"></i>Lista de citas</a></li>
 
+          <?php if ($PERM_OBJ == 1){ ?>
+            <ul class="treeview-menu">
+          <li><a href="../vistas/mostrar_citas_vista.php"><i class="fa fa-list"></i>Lista de citas</a></li>
         </ul>
+           <?php } ?>
+        
       </li>
         </a>
       </li>
  <!-- Titulo de Seguridad -->
- <li class="treeview">
+
+    <?php if ($PERM_OBJ == 1){ ?>
+      <li class="treeview">
         <a href="#">
           <i class="glyphicon glyphicon-lock"></i>
           <span>Seguridad</span>
@@ -177,12 +212,15 @@ $resultado2->execute(array(":id"=>NULL,":usuc"=>$_SESSION["id_us"],":objeto"=>11
         </a>
         <!-- subtitulos de Seguridad -->
         <ul class="treeview-menu">
+        <li><a href="../vistas/insertar_permisos_vista.php"><i class="fa fa-list"></i>Añadir Permisos</a></li>
           <li><a href="../vistas/mostrar_parametros_vista.php"><i class="fa fa-list"></i>Lista de Parámetros</a></li>
           <li><a href="../vistas/mostrar_roles_vista.php"><i class="fa fa-list"></i>Lista de Roles</a></li>
           <li><a href="#"><i class="glyphicon glyphicon-cloud-upload"></i>Backup</a></li>
           <li><a href="../vistas/bitacora_vista.php"><i class="fa fa-list"></i>Bitácora</a></li>
         </ul>
       </li>
+    <?php } ?>
+
    
     </section>
     <!-- /.sidebar -->
@@ -209,10 +247,12 @@ $resultado2->execute(array(":id"=>NULL,":usuc"=>$_SESSION["id_us"],":objeto"=>11
             </div>
             <!--llamar funciones-->
             <div class="box-body">
-           <div>
-
+            <?php if ($INSERTAR == 1){ ?>
+              <div>
              <a href="../vistas/insertar_pacientes_vista.php" class="btn bg-blue btn-flat margin">AGREGAR PACIENTE <i class="fa fa-plus" aria-hidden="true"></i> </a>
            </div>
+            <?php } ?>
+           
               <table id="example1" class="table table-bordered table-striped">
                 <thead>
                 <tr>
@@ -249,14 +289,23 @@ $resultado2->execute(array(":id"=>NULL,":usuc"=>$_SESSION["id_us"],":objeto"=>11
                  <td><?php echo $fila['PER_APELLIDOS']?></td>
                 
                  <td>
-                 <a href='../modelos/mostrar_pacientes_modelo.php?id=<?php echo $fila["PER_CODIGO"]?>' class="btn bg-blue btn-flat margin">
+                 <?php if ($CONSULTAR == 1){ ?>
+
+                  <a href='../modelos/mostrar_pacientes_modelo.php?id=<?php echo $fila["PER_CODIGO"]?>' class="btn bg-blue btn-flat margin">
                  <i class='fa fa-eye'></i></a> 
+                <?php } ?>
+                 
+                <?php if ($ACTUALIZAR == 1){ ?>
 
-					       <a href='../modelos/editar_pacientes_modelo.php?id=<?php echo $fila["PER_CODIGO"]?>' class="btn bg-orange btn-flat margin">
+                  <a href='../modelos/editar_pacientes_modelo.php?id=<?php echo $fila["PER_CODIGO"]?>' class="btn bg-orange btn-flat margin">
                  <i class='fa fa-pencil'></i></a>
-
-                 <a href='../modelos/eliminar_pacientes_modelo.php?id=<?php echo $fila["PER_CODIGO"]?>' class="btn btne bg-maroon bnt-flat margin">
+                 <?php } ?>
+					       
+                 <?php if ($ELIMINAR == 1){ ?>
+                  <a href='../modelos/eliminar_pacientes_modelo.php?id=<?php echo $fila["PER_CODIGO"]?>' class="btn btne bg-maroon bnt-flat margin">
 					       <i class='fa fa-trash'></i></a> 
+                 <?php } ?>
+                 
                  </td>
                  <td><?php echo $fila['PER_PASAPORTE']?></td>
                  <td><?php echo $fila['PER_FECHA_NACIMIENTO']?></td>
